Fixing issue with NCBI efetch. History is not available for this command anymore.

+++ UTILS.dir/retrieve_proteins.sh
@@ -2,25 +2,29 @@
 
 family=$1
 txid=$2
-report_file=$3
 
 
 ###################################
 ## Retrieving sequence from Genbank
-echo "Retrieving representatives of virus family: $txid" 
+echo "Retrieving representatives of virus family $family, TaxID $txid ..." 
 
-`../../UTILS.dir/./fetch_genomes_based_on_taxid.pl $txid` #adicionei o ./ e retirei ~gustavo/devel/robots/robot_genbank/scripts/
+../../UTILS.dir/fetch_genomes_based_on_taxid.pl $txid
 
 # Split genbank file
 # for i in *.gb; do split_genbank.pl $i; done
-`../../UTILS.dir/./split_genbank.pl $txid.gb` #adicionei o ./
+../../UTILS.dir/split_genbank.pl $txid.gb
 
 # Number of genbank files
 num_gb_files=`ls -l *.gb | grep -v original.gb | wc | awk '{print $1}'`
-echo "    Number of sequences retrieved from genbank: " $num_gb_files >> $report_file
+echo "    Number of sequences retrieved from genbank: $num_gb_files" 
+
 
 # List organisms and definition associated to each genbank file
-rm gb_description.txt
+if [ -f gb_description.txt ]
+then 
+    rm gb_description.txt
+fi
+
 for gb_seq in `ls -l *.gb | grep -v original.gb | awk '{print $9}'`
  do org=`awk '$1 == "ORGANISM" {print $0}' $gb_seq | sed 's/ORGANISM//' | sed 's/^ \+//' | sed 's/ /_/g'`
  def=`awk '$1 == "DEFINITION" {print $0}' $gb_seq | sed 's/DEFINITION//' | sed 's/^ \+//'`
@@ -39,7 +43,7 @@ done
 
 # Number of species
 num_species=`awk '{print $2}' gb_description.txt | sort | uniq | wc | awk '{print $1}'`
-echo "    Number of species: " $num_species >> $report_file
+echo "    Number of species: " $num_species
 
 
 ##########################################################
@@ -88,11 +92,11 @@ for i in *.fasta; do transeq $i $i.prot; done
 cat *.prot > all.prot.fas
 
 # Remove "*" STOP codons from each sequence
-`../../UTILS.dir/./fasta2line all.prot.fas | perl -lane '$seq = $F[0]; $seq =~ s/\*$//; print $seq . "\t" . $F[1]' | ../../UTILS.dir/./line2fasta > tmp` ## Adicionei o ./
+../../UTILS.dir/./fasta2line all.prot.fas | perl -lane '$seq = $F[0]; $seq =~ s/\*$//; print $seq . "\t" . $F[1]' | ../../UTILS.dir/./line2fasta > tmp
 mv tmp all.prot.fas
 
 
 # Number of CDS and mat_peptides
 num_CDS_mat=`grep -c ">" all.prot.fas`
-echo "    Number of features: $num_CDS_mat" >> $report_file
+echo "    Number of features: $num_CDS_mat"
